# docs/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "HappyTailz API",
    "version": "1.0.0",
    "description": "Local Swagger documentation for HappyTailz backend"
  },
  "servers": [
    { "url": "http://localhost:3000", "description": "Local" },
    { "url": "https://happytailz-9qf9.onrender.com/", "description": "Render" }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "Pet": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "ownerId": { "type": "string" },
          "name": { "type": "string" },
          "species": { "type": "string" },
          "breed": { "type": "string" },
          "age": { "type": "number" },
          "gender": { "type": "string" },
          "imageUrl": { "type": "string", "format": "uri", "description": "Direct URL of the pet image" }
        }
      },
...
    }
  },
...
  "paths": {
...
    "/api/dashboard": {
      "get": {
        "summary": "User dashboard",
        "description": "Shows user's latest plan with isActive flag, pets summary, upcoming walking requests, and completed count.",
        "tags": ["Dashboard"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "hasPlan": { "type": "boolean" },
                    "plan": { "type": "object", "properties": { "isActive": { "type": "boolean" } } },
                    "pets": { "type": "object", "properties": { "count": { "type": "integer" }, "items": { "type": "array", "items": { "$ref": "#/components/schemas/Pet" } } } },
                    "walkingRequests": { "type": "object", "properties": { "upcoming": { "type": "array" }, "completedCount": { "type": "integer" } } }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/admin/breeding": {
      "get": { "summary": "List breeding", "tags": ["Admin - Breeding"], "security": [{ "bearerAuth": [] }], "responses": { "200": { "description": "OK" } } },
      "post": {
        "summary": "Create breeding",
        "tags": ["Admin - Breeding"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": ["name"],
                "properties": {
                  "name": { "type": "string" },
                  "description": { "type": "string" },
                  "active": { "type": "boolean" },
                  "image": { "type": "string", "format": "binary" },
                  "imageUrl": { "type": "string", "format": "uri" }
                }
              }
            },
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name"],
                "properties": {
                  "name": { "type": "string" },
                  "description": { "type": "string" },
                  "active": { "type": "boolean" },
                  "imageUrl": { "type": "string", "format": "uri" }
                }
              }
            }
          }
        },
        "responses": { "201": { "description": "Created" } }
      }
    },
    "/api/admin/breeding/{id}": {
      "parameters": [
        { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
      ],
      "put": {
        "summary": "Update breeding",
        "tags": ["Admin - Breeding"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string" },
                  "description": { "type": "string" },
                  "active": { "type": "boolean" },
                  "image": { "type": "string", "format": "binary" },
                  "imageUrl": { "type": "string", "format": "uri" }
                }
              }
            },
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string" },
                  "description": { "type": "string" },
                  "active": { "type": "boolean" },
                  "imageUrl": { "type": "string", "format": "uri" }
                }
              }
            }
          }
        },
        "responses": { "200": { "description": "OK" }, "404": { "description": "Not Found" } }
      },
      "delete": { "summary": "Delete breeding", "tags": ["Admin - Breeding"], "security": [{ "bearerAuth": [] }], "responses": { "200": { "description": "Deleted" }, "404": { "description": "Not Found" } } }
    },
    "/api/admin/hospitals": {
      "get": { "summary": "List hospitals", "tags": ["Admin - Hospitals"], "security": [{ "bearerAuth": [] }], "responses": { "200": { "description": "OK" } } },
      "post": {
        "summary": "Create hospital",
        "tags": ["Admin - Hospitals"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": ["name", "address"],
                "properties": {
                  "name": { "type": "string" },
                  "address": { "type": "string" },
                  "phone": { "type": "string" },
                  "services": { "type": "array", "items": { "type": "string" } },
                  "active": { "type": "boolean" },
                  "image": { "type": "string", "format": "binary" },
                  "imageUrl": { "type": "string", "format": "uri" }
                }
              }
            },
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name", "address"],
                "properties": {
                  "name": { "type": "string" },
                  "address": { "type": "string" },
                  "phone": { "type": "string" },
                  "services": { "type": "array", "items": { "type": "string" } },
                  "active": { "type": "boolean" },
                  "imageUrl": { "type": "string", "format": "uri" }
                }
              }
            }
          }
        },
        "responses": { "201": { "description": "Created" } }
      }
    },
    "/api/admin/hospitals/{id}": {
      "parameters": [
        { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
      ],
      "put": {
        "summary": "Update hospital",
        "tags": ["Admin - Hospitals"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string" },
                  "address": { "type": "string" },
                  "phone": { "type": "string" },
                  "services": { "type": "array", "items": { "type": "string" } },
                  "active": { "type": "boolean" },
                  "image": { "type": "string", "format": "binary" },
                  "imageUrl": { "type": "string", "format": "uri" }
                }
              }
            },
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string" },
                  "address": { "type": "string" },
                  "phone": { "type": "string" },
                  "services": { "type": "array", "items": { "type": "string" } },
                  "active": { "type": "boolean" },
                  "imageUrl": { "type": "string", "format": "uri" }
                }
              }
            }
          }
        },
        "responses": { "200": { "description": "OK" }, "404": { "description": "Not Found" } }
      },
      "delete": { "summary": "Delete hospital", "tags": ["Admin - Hospitals"], "security": [{ "bearerAuth": [] }], "responses": { "200": { "description": "Deleted" }, "404": { "description": "Not Found" } } }
    },
    "/api/admin/grooming": {
      "get": { "summary": "List grooming", "tags": ["Admin - Grooming"], "security": [{ "bearerAuth": [] }], "responses": { "200": { "description": "OK" } } },
      "post": {
        "summary": "Create grooming",
        "tags": ["Admin - Grooming"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": ["name", "price"],
                "properties": {
                  "name": { "type": "string" },
                  "services": { "type": "array", "items": { "type": "string" } },
                  "price": { "type": "number" },
                  "durationMinutes": { "type": "integer" },
                  "active": { "type": "boolean" },
                  "image": { "type": "string", "format": "binary" },
                  "imageUrl": { "type": "string", "format": "uri" }
                }
              }
            },
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name", "price"],
                "properties": {
                  "name": { "type": "string" },
                  "services": { "type": "array", "items": { "type": "string" } },
                  "price": { "type": "number" },
                  "durationMinutes": { "type": "integer" },
                  "active": { "type": "boolean" },
                  "imageUrl": { "type": "string", "format": "uri" }
                }
              }
            }
          }
        },
        "responses": { "201": { "description": "Created" } }
      }
    },
    "/api/admin/grooming/{id}": {
      "parameters": [
        { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
      ],
      "put": {
        "summary": "Update grooming",
        "tags": ["Admin - Grooming"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string" },
                  "services": { "type": "array", "items": { "type": "string" } },
                  "price": { "type": "number" },
                  "durationMinutes": { "type": "integer" },
                  "active": { "type": "boolean" },
                  "image": { "type": "string", "format": "binary" },
                  "imageUrl": { "type": "string", "format": "uri" }
                }
              }
            },
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string" },
                  "services": { "type": "array", "items": { "type": "string" } },
                  "price": { "type": "number" },
                  "durationMinutes": { "type": "integer" },
                  "active": { "type": "boolean" },
                  "imageUrl": { "type": "string", "format": "uri" }
                }
              }
            }
          }
        },
        "responses": { "200": { "description": "OK" }, "404": { "description": "Not Found" } }
      },
      "delete": { "summary": "Delete grooming", "tags": ["Admin - Grooming"], "security": [{ "bearerAuth": [] }], "responses": { "200": { "description": "Deleted" }, "404": { "description": "Not Found" } } }
    },
    "/api/admin/plans": {
      "get": { "summary": "List plans", "tags": ["Admin - Plans"], "responses": { "200": { "description": "OK" } } },
      "post": {
        "summary": "Create plan",
        "tags": ["Admin - Plans"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name", "durationDays", "price"],
                "properties": {
                  "name": { "type": "string", "enum": ["1 day", "1 week", "1 month"] },
                  "durationDays": { "type": "integer" },
                  "price": { "type": "number" },
                  "description": { "type": "string" },
                  "active": { "type": "boolean" }
                }
              }
            }
          }
        },
        "responses": { "201": { "description": "Created" } }
      }
    },
    "/api/admin/plans/{id}": {
      "parameters": [
        { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
      ],
      "put": {
        "summary": "Update plan",
        "tags": ["Admin - Plans"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string", "enum": ["1 day", "1 week", "1 month"] },
                  "durationDays": { "type": "integer" },
                  "price": { "type": "number" },
                  "description": { "type": "string" },
                  "active": { "type": "boolean" }
                }
              }
            }
          }
        },
        "responses": { "200": { "description": "OK" }, "404": { "description": "Not Found" } }
      },
      "delete": { "summary": "Delete plan", "tags": ["Admin - Plans"], "security": [{ "bearerAuth": [] }], "responses": { "200": { "description": "Deleted" }, "404": { "description": "Not Found" } } }
    },
...
    "/api/admin/dashboard": {
      "get": {
        "summary": "Admin dashboard summary",
        "description": "High-level counts: users by role, pets, active subscriptions, walking requests by status, and recent users.",
        "tags": ["Admin - Dashboard"],
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "description": "OK" } }
      }
    },
    "/api/admin/dashboard/plan-stats": {
      "get": {
        "summary": "Plan subscription stats",
        "description": "Counts how many user plan subscriptions were created in the period, grouped by plan name.",
        "tags": ["Admin - Dashboard"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "period", "in": "query", "schema": { "type": "string", "enum": ["day", "week", "month" ] }, "description": "Defaults to month" }
        ],
        "responses": { "200": { "description": "OK" } }
      }
    }
  }
}